+++ packages/downloader/src/utils/parser/index.ts
@@ -4,7 +4,7 @@ import {
 } from './share'
 import { type Package } from '@dep-mgr/share'
 import { LockFile } from '../../constants'
-import { readLockFile } from '../reader'
+import { readLockFilePath } from '../reader'
 import { parsePackages as npmParsePackages } from './npm'
 import { parsePackages as pnpmParsePackages } from './pnpm'
 import { parsePackages as yarnParsePackages } from './yarn'
@@ -16,10 +16,10 @@ const lockFileParsePackagesMap: Record<LockFile, ParsePackagesFunction> = {
   [LockFile.PnpmYamlLockFile]: pnpmParsePackages
 }
 
-export const parsePackages: RootParsePackagesFunction = (
+export const parsePackages: RootParsePackagesFunction = async (
   lockFile: LockFile
-): Package[] => {
-  const lockFileContent = readLockFile(lockFile)
-  const packages = lockFileParsePackagesMap[lockFile](lockFile, lockFileContent)
+): Promise<Package[]> => {
+  const lockFilePath = readLockFilePath(lockFile)
+  const packages = await lockFileParsePackagesMap[lockFile](lockFile, lockFilePath)
   return packages
 }