+++ packages/downloader/src/utils/parser/npm.ts
@@ -3,9 +3,10 @@ import { type LockFile } from '../../constants'
 import { type ParsePackagesFunction } from './share'
 import { ParserError, ParserErrorCode } from '../../error/parser-error'
 
-export const parsePackages: ParsePackagesFunction = (
+export const parsePackages: ParsePackagesFunction = async (
   lockFile: LockFile,
-  lockFileContent: string
-): Package[] => {
+  lockFilePath: string
+// eslint-disable-next-line @typescript-eslint/require-await
+): Promise<Package[]> => {
   throw new ParserError(ParserErrorCode.NOT_IMPLEMENT)
 }